BaseCommand moved to Base folder
